MedicalAPI_model/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pytesseract
opencv-python
pillow
numpy
